+++ src/SqlArtisan/Internal/SqlPart/Clause/Merge/MergeIntoClause.cs
@@ -1,0 +1,10 @@
+﻿namespace SqlArtisan.Internal;
+
+internal sealed class MergeIntoClause(DbTableBase target) : SqlPart
+{
+    private readonly DbTableBase _target = target;
+
+    internal override void Format(SqlBuildingBuffer buffer) => buffer
+        .Append($"{Keywords.Merge} {Keywords.Into} ")
+        .Append(_target);
+}