+++ src/SqlArtisan/Internal/SqlPart/Clause/Merge/MergeUsingClause.cs
@@ -1,0 +1,10 @@
+﻿namespace SqlArtisan.Internal;
+
+internal sealed class MergeUsingClause(TableReference source) : SqlPart
+{
+    private readonly TableReference _source = source;
+
+    internal override void Format(SqlBuildingBuffer buffer) => buffer
+        .Append($"{Keywords.Using} ")
+        .Append(_source);
+}